# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Asmblyr Core HTTP API",
    "version": "0.0.0",
    "description": "Все статические маршруты Core и их требования доступа. Часть операций имеет только route-only описание; спецификация не является полной схемой для генерации клиента. OAuth protocol и динамические маршруты плагинов описаны отдельно."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Локальный Core"
    }
  ],
  "paths": {
...
    "/permissions": {
      "get": {
        "tags": [
          "permissions"
        ],
        "operationId": "get__permissions",
        "summary": "GET /permissions",
        "description": "Человек: policies/read ИЛИ update; superuser bypass. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/permissions/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/permissions/routes.ts"
      },
      "post": {
        "tags": [
          "permissions"
        ],
        "operationId": "post__permissions",
        "summary": "POST /permissions",
        "description": "Только человек-superuser: изменение состава политик и permissions. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/permissions/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/permissions/routes.ts"
      }
    },
    "/permissions/{id}": {
      "delete": {
        "tags": [
          "permissions"
        ],
        "operationId": "delete__permissions_id",
        "summary": "DELETE /permissions/:id",
        "description": "Только человек-superuser: изменение состава политик и permissions. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/permissions/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/permissions/routes.ts"
      },
      "get": {
        "tags": [
          "permissions"
        ],
        "operationId": "get__permissions_id",
        "summary": "GET /permissions/:id",
        "description": "Человек: policies/read ИЛИ update; superuser bypass. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/permissions/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/permissions/routes.ts"
      },
      "patch": {
        "tags": [
          "permissions"
        ],
        "operationId": "patch__permissions_id",
        "summary": "PATCH /permissions/:id",
        "description": "Только человек-superuser: изменение состава политик и permissions. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/permissions/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/permissions/routes.ts"
      }
    },
    "/permissions/me": {
      "get": {
        "tags": [
          "permissions"
        ],
        "operationId": "get__permissions_me",
        "summary": "GET /permissions/me",
        "description": "Активный principal; только собственные права. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/permissions/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/permissions/routes.ts"
      }
    },
...
    "/settings/plugins": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "get__settings_plugins",
        "summary": "GET /settings/plugins",
        "description": "Человек: plugins/read ИЛИ update; superuser bypass. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/plugins/settings-routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/plugins/settings-routes.ts"
      }
    },
    "/settings/plugins/{namespace}": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "get__settings_plugins_namespace",
        "summary": "GET /settings/plugins/:namespace",
        "description": "Человек: plugins/read ИЛИ update; superuser bypass. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/plugins/settings-routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/plugins/settings-routes.ts"
      },
      "put": {
        "tags": [
          "settings"
        ],
        "operationId": "put__settings_plugins_namespace",
        "summary": "PUT /settings/plugins/:namespace",
        "description": "Человек: plugins/update; superuser bypass. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/plugins/settings-routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/plugins/settings-routes.ts"
      }
    },
...
    "/users/me/preferences": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get__users_me_preferences",
        "summary": "GET /users/me/preferences",
        "description": "Человек, только собственный профиль/ресурс; workspace дополнительно проверяет видимость. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/preferences/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/preferences/routes.ts"
      },
      "patch": {
        "tags": [
          "users"
        ],
        "operationId": "patch__users_me_preferences",
        "summary": "PATCH /users/me/preferences",
        "description": "Человек, только собственный профиль/ресурс; workspace дополнительно проверяет видимость. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/preferences/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/preferences/routes.ts"
      }
    },
...
    "/users/me/table-preferences/{collection}": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get__users_me_table_preferences_collection",
        "summary": "GET /users/me/table-preferences/:collection",
        "description": "Человек, только собственный профиль/ресурс; workspace дополнительно проверяет видимость. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/preferences/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/preferences/routes.ts"
      },
      "patch": {
        "tags": [
          "users"
        ],
        "operationId": "patch__users_me_table_preferences_collection",
        "summary": "PATCH /users/me/table-preferences/:collection",
        "description": "Человек, только собственный профиль/ресурс; workspace дополнительно проверяет видимость. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/preferences/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/preferences/routes.ts"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Непрозрачный access token пользователя или сервиса. Разделы настроек доступны только человеку."
      }
    }
  },
...
}